# device-farm/infra/api/api-spec.yaml
openapi: 3.0.3
info:
  title: Device Farm API
  description: 设备农场管理平台 API 文档
  version: 1.0.0
  contact:
    name: Device Farm Team
    email: device-farm@example.com
servers:
  - url: http://localhost:80
    description: 开发环境
  - url: https://device-farm.example.com
    description: 生产环境
...
paths:
...
  /api/v1/scripts:
    get:
      tags:
        - scripts
      summary: 获取脚本列表
      operationId: listScripts
      parameters:
        - name: language
          in: query
          description: 脚本语言
          schema:
            type: string
            enum: [python, javascript, appium, airtest]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptListResponse'

    post:
      tags:
        - scripts
      summary: 创建脚本
      operationId: createScript
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScriptCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/scripts/{scriptId}:
    get:
      tags:
        - scripts
      summary: 获取脚本详情
      operationId: getScript
      parameters:
        - $ref: '#/components/parameters/scriptId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - scripts
      summary: 更新脚本
      operationId: updateScript
      parameters:
        - $ref: '#/components/parameters/scriptId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScriptUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - scripts
      summary: 删除脚本
      operationId: deleteScript
      parameters:
        - $ref: '#/components/parameters/scriptId'
      responses:
        '204':
          description: 删除成功
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== 测试任务 ====================
  /api/v1/tasks:
    get:
      tags:
        - tasks
      summary: 获取任务列表
      operationId: listTasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
        - name: deviceId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'

    post:
      tags:
        - tasks
      summary: 创建测试任务
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: 设备不可用
  /api/v1/tasks/{taskId}:
    get:
      tags:
        - tasks
      summary: 获取任务详情
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - tasks
      summary: 取消任务
      operationId: cancelTask
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: 任务已取消
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
...
    scriptId:
      name: scriptId
      in: path
      required: true
      description: 脚本ID
      schema:
        type: string
    taskId:
      name: taskId
      in: path
      required: true
      description: 任务ID
      schema:
        type: string
...
  schemas:
...
    Script:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        language:
          type: string
          enum: [python, javascript, appium, airtest]
        content:
          type: string
        version:
          type: integer
        description:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ScriptCreateRequest:
      type: object
      required:
        - name
        - language
      properties:
        name:
          type: string
        language:
          type: string
          enum: [python, javascript, appium, airtest]
        content:
          type: string
        description:
          type: string
    ScriptUpdateRequest:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
        description:
          type: string
    ScriptResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/Script'
    ScriptListResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            total:
              type: integer
            page:
              type: integer
            pageSize:
              type: integer
            items:
              type: array
              items:
                $ref: '#/components/schemas/Script'

    # 任务相关
    Task:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        scriptId:
          type: string
        deviceId:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        priority:
          type: integer
        params:
          type: object
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
    TaskCreateRequest:
      type: object
      required:
        - name
        - scriptId
        - deviceId
      properties:
        name:
          type: string
        scriptId:
          type: string
        deviceId:
          type: string
        priority:
          type: integer
          default: 0
        params:
          type: object
    TaskResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/Task'
    TaskListResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            total:
              type: integer
            page:
              type: integer
            pageSize:
              type: integer
            items:
              type: array
              items:
                $ref: '#/components/schemas/Task'

    # 报告相关
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...